requirements.txt
asgiref==3.3.1
Django==3.1.3
django-environ==0.4.5
psycopg2-binary==2.8.5
djangorestframework==3.12.2
drf-extensions==0.6.0
pytz==2020.4
sqlparse==0.4.1
